test_requirements.txt
django-coverage
coverage
tox
ipdb
